+++ .gitignore
@@ -6,3 +6,4 @@ do.sh
 tests.py
 test2.py
 tests.py
+__pycache__/db_works.cpython-38.pyc
