+++ db_works.py
@@ -6,8 +6,8 @@ import json
 def get_conn_settings_from_json():
     with open("db_credentials.json") as json_conf:
         sql_db_conn = (json.load(json_conf))
+    print("get_conn_settings_from_json done")
     return sql_db_conn
-
 
 # connection to db
 def db_connect():
@@ -17,8 +17,8 @@ def db_connect():
                                    host=sql_db_conn["host"],
                                    database=sql_db_conn["database"])
     cursor = cnxn.cursor()
+    print("db_connect done")
     return cursor, cnxn
-
 
 # get table names
 def db_tables():
@@ -26,4 +26,5 @@ def db_tables():
     db_schema_name = sql_db_conn["db_schema_name"]  # schema_name
     db_table_name = sql_db_conn["db_table_name"]  # table_name
     db_settings_table_name = sql_db_conn["db_settings_table_name"]  # settings table name
+    print("db_tables done")
     return db_schema_name, db_table_name, db_settings_table_name
